Refactored the name of the styled components folders for Lists and Divs.
