+++ src/style-components/index.ts
@@ -1,10 +1,10 @@
 export * from './Button';
 export * from './Header';
 export * from './Svg';
-export * from './Wrapper';
+export * from './Div';
 export * from './Link';
 export * from './Paragraph';
 export * from './Section';
 export * from './Svg';
-export * from './Table';
+export * from './List';
 export * from './Span'
